f2c: Changed return type of r_imag(complex *) from double to float

+++ F2CLIBS/libf2c/r_imag.c
@@ -4,9 +4,9 @@ extern "C" {
 #endif
 
 #ifdef KR_headers
-double r_imag(z) complex *z;
+float r_imag(z) complex *z;
 #else
-double r_imag(complex *z)
+float r_imag(complex *z)
 #endif
 {
 return(z->i);
